requirements.txt
# Mac Agent 의존성

# AI/ML 라이브러리
autogen-agentchat>=0.2.0
autogen-core>=0.2.0
autogen-ext[openai]>=0.2.0
openai>=1.0.0
pydantic>=2.0.0

# 환경 설정
python-dotenv>=1.0.0

# macOS 시스템 통합
pyobjc-framework-EventKit>=10.0
pyobjc-core>=10.0

# 유틸리티
python-dateutil>=2.8.0 